+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.4] - 2018-02-26
+### Added
+- Added stats tokens
+- Added temporary button tokens
+- Added icon-filter, icon-mail-multi, icon-clear-formatting
+
 ## [1.1.3] - 2018-02-22
 ### Changed
 - Updated paths in readme.